+++ priv/repo/migrations/20240217041933_create_bookings.exs
@@ -5,7 +5,7 @@ defmodule VacationNest.Repo.Migrations.CreateBookings do
     create table(:bookings, primary_key: false) do
       add :id, :binary_id, primary_key: true
       add :user_id, references(:users, type: :uuid)
-      add :total_amount, :float
+      add :total_amount, :integer
       add :check_in_day, :date
       add :check_out_day, :date
       add :status, :string
